backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML Dependencies
openai==1.3.5
langchain==0.0.339
langchain-openai==0.0.2
sentence-transformers==2.2.2
pinecone-client==2.2.4
tiktoken==0.5.1

# Document Processing
pypdf2==3.0.1
python-docx==0.8.11
python-pptx==0.6.21
pillow==10.1.0
pytesseract==0.3.10
unstructured==0.11.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Async & Caching
redis==5.0.1
celery==5.3.4
httpx==0.25.2

# Utilities
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
pre-commit==3.5.0 